refactor: move timestimp cast to common
so that connector doesn't depend on expr
