+++ src/common/Cargo.toml
@@ -71,6 +71,7 @@ serde_default = "0.1"
 serde_json = "1"
 serde_with = "2"
 smallbitset = "0.6.1"
+speedate = "0.7.0"
 static_assertions = "1"
 strum = "0.24"
 strum_macros = "0.24"
