+++ src/common/src/lib.rs
@@ -36,6 +36,7 @@
 #![feature(iterator_try_collect)]
 #![feature(round_ties_even)]
 #![feature(iter_order_by)]
+#![feature(exclusive_range_pattern)]
 
 #[macro_use]
 pub mod jemalloc;
@@ -47,6 +48,7 @@ pub mod array;
 pub mod util;
 pub mod buffer;
 pub mod cache;
+pub mod cast;
 pub mod catalog;
 pub mod config;
 pub mod constants;
