+++ src/connector/src/parser/csv_parser.rs
@@ -16,12 +16,10 @@ use std::str::FromStr;
 
 use anyhow::anyhow;
 use futures_async_stream::try_stream;
+use risingwave_common::cast::{str_to_date, str_to_timestamp, str_with_time_zone_to_timestamptz};
 use risingwave_common::error::ErrorCode::{InternalError, ProtocolError};
 use risingwave_common::error::{Result, RwError};
 use risingwave_common::types::{Datum, Decimal, ScalarImpl};
-use risingwave_expr::vector_op::cast::{
-    str_to_date, str_to_timestamp, str_with_time_zone_to_timestamptz,
-};
 
 use crate::impl_common_parser_logic;
 use crate::parser::{SourceStreamChunkRowWriter, WriteGuard};
