+++ src/connector/src/parser/json_parser.rs
@@ -141,11 +141,11 @@ mod tests {
 
     use itertools::Itertools;
     use risingwave_common::array::{Op, StructValue};
+    use risingwave_common::cast::{str_to_date, str_to_timestamp};
     use risingwave_common::catalog::ColumnDesc;
     use risingwave_common::row::Row;
     use risingwave_common::test_prelude::StreamChunkTestExt;
     use risingwave_common::types::{DataType, Decimal, ScalarImpl, ToOwnedDatum};
-    use risingwave_expr::vector_op::cast::{str_to_date, str_to_timestamp};
 
     use crate::common::UpsertMessage;
     use crate::parser::{JsonParser, SourceColumnDesc, SourceStreamChunkBuilder};
