+++ src/expr/src/vector_op/cast.rs
@@ -16,11 +16,13 @@ use std::any::type_name;
 use std::fmt::Write;
 use std::str::FromStr;
 
-use chrono::{DateTime, NaiveDate, NaiveDateTime, NaiveTime, Utc};
 use futures_util::FutureExt;
 use itertools::Itertools;
 use risingwave_common::array::{
     ListArray, ListRef, ListValue, StructArray, StructRef, StructValue, Utf8Array,
+};
+use risingwave_common::cast::{
+    parse_naive_date, parse_naive_datetime, parse_naive_time, str_with_time_zone_to_timestamptz,
 };
 use risingwave_common::row::OwnedRow;
 use risingwave_common::types::{
@@ -30,7 +32,6 @@ use risingwave_common::types::{
 use risingwave_common::util::iter_util::ZipEqFast;
 use risingwave_expr_macro::{build_function, function};
 use risingwave_pb::expr::expr_node::PbType;
-use speedate::{Date as SpeedDate, DateTime as SpeedDateTime, Time as SpeedTime};
 
 use crate::expr::template::UnaryExpression;
 use crate::expr::{build, BoxedExpression, Expression, InputRefExpression};
@@ -43,150 +44,28 @@ const TRUE_BOOL_LITERALS: [&str; 9] = ["true", "tru", "tr", "t", "on", "1", "yes
 const FALSE_BOOL_LITERALS: [&str; 10] = [
     "false", "fals", "fal", "fa", "f", "off", "of", "0", "no", "n",
 ];
-const ERROR_INT_TO_TIMESTAMP: &str = "Can't cast negative integer to timestamp";
-const PARSE_ERROR_STR_WITH_TIME_ZONE_TO_TIMESTAMPTZ: &str = concat!(
-    "Can't cast string to timestamp with time zone (expected format is YYYY-MM-DD HH:MM:SS[.D+{up to 6 digits}] followed by +hh:mm or literal Z)"
-    , "\nFor example: '2021-04-01 00:00:00+00:00'"
-);
-const PARSE_ERROR_STR_TO_TIMESTAMP: &str = "Can't cast string to timestamp (expected format is YYYY-MM-DD HH:MM:SS[.D+{up to 6 digits}] or YYYY-MM-DD HH:MM or YYYY-MM-DD or ISO 8601 format)";
-const PARSE_ERROR_STR_TO_TIME: &str =
-    "Can't cast string to time (expected format is HH:MM:SS[.D+{up to 6 digits}] or HH:MM)";
-const PARSE_ERROR_STR_TO_DATE: &str = "Can't cast string to date (expected format is YYYY-MM-DD)";
+
 const PARSE_ERROR_STR_TO_BYTEA: &str = "Invalid Bytea syntax";
 
 #[function("cast(varchar) -> date")]
 pub fn str_to_date(elem: &str) -> Result<Date> {
-    Ok(Date::new(parse_naive_date(elem)?))
+    Ok(Date::new(
+        parse_naive_date(elem).map_err(|err| ExprError::Parse(err.into()))?,
+    ))
 }
 
 #[function("cast(varchar) -> time")]
 pub fn str_to_time(elem: &str) -> Result<Time> {
-    Ok(Time::new(parse_naive_time(elem)?))
+    Ok(Time::new(
+        parse_naive_time(elem).map_err(|err| ExprError::Parse(err.into()))?,
+    ))
 }
 
 #[function("cast(varchar) -> timestamp")]
 pub fn str_to_timestamp(elem: &str) -> Result<Timestamp> {
-    Ok(Timestamp::new(parse_naive_datetime(elem)?))
-}
-
-#[inline]
-fn parse_naive_datetime(s: &str) -> Result<NaiveDateTime> {
-    if let Ok(res) = SpeedDateTime::parse_str(s) {
-        Ok(Date::from_ymd_uncheck(
-            res.date.year as i32,
-            res.date.month as u32,
-            res.date.day as u32,
-        )
-        .and_hms_micro_uncheck(
-            res.time.hour as u32,
-            res.time.minute as u32,
-            res.time.second as u32,
-            res.time.microsecond,
-        )
-        .0)
-    } else {
-        let res = SpeedDate::parse_str(s)
-            .map_err(|_| ExprError::Parse(PARSE_ERROR_STR_TO_TIMESTAMP.into()))?;
-        Ok(
-            Date::from_ymd_uncheck(res.year as i32, res.month as u32, res.day as u32)
-                .and_hms_micro_uncheck(0, 0, 0, 0)
-                .0,
-        )
-    }
-}
-
-/// Converts UNIX epoch time to timestamp.
-///
-/// The input UNIX epoch time is interpreted as follows:
-///
-/// - [0, 1e11) are assumed to be in seconds.
-/// - [1e11, 1e14) are assumed to be in milliseconds.
-/// - [1e14, 1e17) are assumed to be in microseconds.
-/// - [1e17, upper) are assumed to be in nanoseconds.
-///
-/// This would cause no problem for timestamp in [1973-03-03 09:46:40, 5138-11-16 09:46:40).
-///
-/// # Example
-/// ```
-/// # use risingwave_expr::vector_op::cast::i64_to_timestamp;
-/// assert_eq!(
-///     i64_to_timestamp(1_666_666_666).unwrap().to_string(),
-///     "2022-10-25 02:57:46"
-/// );
-/// assert_eq!(
-///     i64_to_timestamp(1_666_666_666_666).unwrap().to_string(),
-///     "2022-10-25 02:57:46.666"
-/// );
-/// assert_eq!(
-///     i64_to_timestamp(1_666_666_666_666_666).unwrap().to_string(),
-///     "2022-10-25 02:57:46.666666"
-/// );
-/// assert_eq!(
-///     i64_to_timestamp(1_666_666_666_666_666_666)
-///         .unwrap()
-///         .to_string(),
-///     // note that we only support microseconds precision
-///     "2022-10-25 02:57:46.666666"
-/// );
-/// ```
-#[inline]
-pub fn i64_to_timestamp(t: i64) -> Result<Timestamp> {
-    let us = i64_to_timestamptz(t)?;
-    Ok(Timestamp::from_timestamp_uncheck(
-        us / 1_000_000,
-        (us % 1_000_000) as u32 * 1000,
+    Ok(Timestamp::new(
+        parse_naive_datetime(elem).map_err(|err| ExprError::Parse(err.into()))?,
     ))
-}
-
-#[inline]
-fn parse_naive_date(s: &str) -> Result<NaiveDate> {
-    let res =
-        SpeedDate::parse_str(s).map_err(|_| ExprError::Parse(PARSE_ERROR_STR_TO_DATE.into()))?;
-    Ok(Date::from_ymd_uncheck(res.year as i32, res.month as u32, res.day as u32).0)
-}
-
-#[inline]
-fn parse_naive_time(s: &str) -> Result<NaiveTime> {
-    let res =
-        SpeedTime::parse_str(s).map_err(|_| ExprError::Parse(PARSE_ERROR_STR_TO_TIME.into()))?;
-    Ok(Time::from_hms_micro_uncheck(
-        res.hour as u32,
-        res.minute as u32,
-        res.second as u32,
-        res.microsecond,
-    )
-    .0)
-}
-
-#[inline(always)]
-pub fn str_with_time_zone_to_timestamptz(elem: &str) -> Result<i64> {
-    elem.parse::<DateTime<Utc>>()
-        .map(|ret| ret.timestamp_micros())
-        .map_err(|_| ExprError::Parse(PARSE_ERROR_STR_WITH_TIME_ZONE_TO_TIMESTAMPTZ.into()))
-}
-
-/// Converts UNIX epoch time to timestamp in microseconds.
-///
-/// The input UNIX epoch time is interpreted as follows:
-///
-/// - [0, 1e11) are assumed to be in seconds.
-/// - [1e11, 1e14) are assumed to be in milliseconds.
-/// - [1e14, 1e17) are assumed to be in microseconds.
-/// - [1e17, upper) are assumed to be in nanoseconds.
-///
-/// This would cause no problem for timestamp in [1973-03-03 09:46:40, 5138-11-16 09:46:40).
-#[inline]
-pub fn i64_to_timestamptz(t: i64) -> Result<i64> {
-    const E11: i64 = 100_000_000_000;
-    const E14: i64 = 100_000_000_000_000;
-    const E17: i64 = 100_000_000_000_000_000;
-    match t {
-        0..E11 => Ok(t * 1_000_000), // s
-        E11..E14 => Ok(t * 1_000),   // ms
-        E14..E17 => Ok(t),           // us
-        E17.. => Ok(t / 1_000),      // ns
-        _ => Err(ExprError::Parse(ERROR_INT_TO_TIMESTAMP.into())),
-    }
 }
 
 #[function("cast(varchar) -> bytea")]
@@ -451,7 +330,9 @@ pub fn literal_parsing(
         DataType::Timestamp => str_to_timestamp(s)?.into(),
         // We only handle the case with timezone here, and leave the implicit session timezone case
         // for later phase.
-        DataType::Timestamptz => str_with_time_zone_to_timestamptz(s)?.into(),
+        DataType::Timestamptz => str_with_time_zone_to_timestamptz(s)
+            .map_err(|err| ExprError::Parse(err.into()))?
+            .into(),
         DataType::Time => str_to_time(s)?.into(),
         DataType::Interval => str_parse::<Interval>(s)?.into(),
         // Not processing list or struct literal right now. Leave it for later phase (normal backend
@@ -639,47 +520,10 @@ fn struct_cast(
 
 #[cfg(test)]
 mod tests {
+    use chrono::NaiveDateTime;
     use risingwave_common::types::Scalar;
 
     use super::*;
-
-    #[test]
-    fn parse_str() {
-        assert_eq!(
-            str_with_time_zone_to_timestamptz("2022-08-03 10:34:02Z").unwrap(),
-            str_with_time_zone_to_timestamptz("2022-08-03 02:34:02-08:00").unwrap()
-        );
-        str_to_timestamp("1999-01-08 04:02").unwrap();
-        str_to_timestamp("1999-01-08 04:05:06").unwrap();
-        assert_eq!(
-            str_to_timestamp("2022-08-03T10:34:02Z").unwrap(),
-            str_to_timestamp("2022-08-03 10:34:02").unwrap()
-        );
-        str_to_date("1999-01-08").unwrap();
-        str_to_time("04:05").unwrap();
-        str_to_time("04:05:06").unwrap();
-
-        assert_eq!(
-            str_with_time_zone_to_timestamptz("1999-01-08 04:05:06")
-                .unwrap_err()
-                .to_string(),
-            ExprError::Parse(PARSE_ERROR_STR_WITH_TIME_ZONE_TO_TIMESTAMPTZ.into()).to_string()
-        );
-        assert_eq!(
-            str_to_timestamp("1999-01-08 04:05:06AA")
-                .unwrap_err()
-                .to_string(),
-            ExprError::Parse(PARSE_ERROR_STR_TO_TIMESTAMP.into()).to_string()
-        );
-        assert_eq!(
-            str_to_date("1999-01-08AA").unwrap_err().to_string(),
-            "Parse error: Can't cast string to date (expected format is YYYY-MM-DD)".to_string()
-        );
-        assert_eq!(
-            str_to_time("AA04:05:06").unwrap_err().to_string(),
-            ExprError::Parse(PARSE_ERROR_STR_TO_TIME.into()).to_string()
-        );
-    }
 
     #[test]
     fn integer_cast_to_bool() {
