+++ src/expr/src/vector_op/timestamptz.rs
@@ -16,10 +16,10 @@ use std::fmt::Write;
 
 use chrono::{TimeZone, Utc};
 use chrono_tz::Tz;
+use risingwave_common::cast::{str_to_timestamp, str_with_time_zone_to_timestamptz};
 use risingwave_common::types::{IntoOrdered, Timestamp, F64};
 use risingwave_expr_macro::function;
 
-use crate::vector_op::cast::{str_to_timestamp, str_with_time_zone_to_timestamptz};
 use crate::{ExprError, Result};
 
 /// Just a wrapper to reuse the `map_err` logic.
@@ -85,8 +85,12 @@ pub fn timestamptz_to_string(elem: i64, time_zone: &str, writer: &mut dyn Write)
 // timestamp and then adjusts it with the session timezone.
 #[function("cast_with_time_zone(varchar, varchar) -> timestamptz")]
 pub fn str_to_timestamptz(elem: &str, time_zone: &str) -> Result<i64> {
-    str_with_time_zone_to_timestamptz(elem)
-        .or_else(|_| timestamp_at_time_zone(str_to_timestamp(elem)?, time_zone))
+    str_with_time_zone_to_timestamptz(elem).or_else(|_| {
+        timestamp_at_time_zone(
+            str_to_timestamp(elem).map_err(|err| ExprError::Parse(err.into()))?,
+            time_zone,
+        )
+    })
 }
 
 #[function("at_time_zone(timestamptz, varchar) -> timestamp")]
